// scripts/types/database/public/EspeceProtegeeReference.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Identifier type for public.espece_protegee_reference */
export type EspeceProtegeeReferenceCdRef = string & { __brand: 'public.espece_protegee_reference' };

/** Represents the table public.espece_protegee_reference */
export default interface EspeceProtegeeReference {
  cd_ref: EspeceProtegeeReferenceCdRef;

  classification: string;

  noms_scientifiques: string[];

  noms_vernaculaires: string[];

  cd_type_statuts: string[];
}

/** Represents the initializer for the table public.espece_protegee_reference */
export interface EspeceProtegeeReferenceInitializer {
  cd_ref: EspeceProtegeeReferenceCdRef;

  classification: string;

  /** Default value: '{}'::text[] */
  noms_scientifiques?: string[];

  /** Default value: '{}'::text[] */
  noms_vernaculaires?: string[];

  /** Default value: '{}'::text[] */
  cd_type_statuts?: string[];
}

/** Represents the mutator for the table public.espece_protegee_reference */
export interface EspeceProtegeeReferenceMutator {
  cd_ref?: EspeceProtegeeReferenceCdRef;

  classification?: string;

  noms_scientifiques?: string[];

  noms_vernaculaires?: string[];

  cd_type_statuts?: string[];
}